+++ components/app/explore/DatasetList.js
@@ -1,13 +1,14 @@
 import React from 'react';
-import find from 'lodash/find';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
+
+import { connect } from 'react-redux';
 
 // Components
 import DatasetWidget from 'components/app/explore/DatasetWidget';
 
 function DatasetList(props) {
-  const { list, mode, showActions, showFavorite, favorites } = props;
+  const { list, mode, showActions, showFavorite, user } = props;
 
   const newClassName = classNames({
     column: true,
@@ -24,15 +25,13 @@ function DatasetList(props) {
           (<div className={newClassName} key={dataset.id}>
             <DatasetWidget
               dataset={dataset}
-              favorite={favorites && favorites.find(elem =>
-                elem.attributes.resourceId === dataset.id)}
-              widget={find(dataset.attributes.widget, { attributes: { default: true } })}
-              layer={find(dataset.attributes.layer, { attributes: { default: true } })}
+              favourite={user.favourites.find(f => f.attributes.resourceId === dataset.id)}
+              widget={dataset.attributes.widget.find(w => w.attributes.default)}
+              layer={dataset.attributes.layer.find(l => l.attributes.default)}
               mode={mode}
               showActions={showActions}
               showFavorite={showFavorite}
               onTagSelected={tag => props.onTagSelected(tag)}
-              onFavoriteRemoved={favorite => props.onFavoriteRemoved(favorite)}
             />
           </div>)
         )}
@@ -43,7 +42,7 @@ function DatasetList(props) {
 
 DatasetList.propTypes = {
   list: PropTypes.array,
-  favorites: PropTypes.array,
+  user: PropTypes.object,
   mode: PropTypes.string,
   showActions: PropTypes.bool.isRequired,
   showFavorite: PropTypes.bool.isRequired,
@@ -53,4 +52,8 @@ DatasetList.propTypes = {
   onFavoriteRemoved: PropTypes.func // eslint-disable-line
 };
 
-export default DatasetList;
+export default connect(
+  state => ({
+    user: state.user
+  })
+)(DatasetList);